Add adapters' documentation and change the adapters' file structures
