+++ src/adapters/fabric/constant.js
