+++ src/adapters/sawtooth/Application/SimpleBatchBuilder.js
